add on exception handling

+++ src/main/java/com/example/demo/connection/StudentController.java
@@ -1,9 +1,11 @@
 package com.example.demo.connection;
 
 import java.util.List;
+import java.util.*;
 
 import javax.validation.Valid;
 
+import org.apache.tomcat.util.file.ConfigurationSource.Resource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,24 +19,41 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-
+import com.example.demo.exception.CustomizedResponseEntityExceptionHandler;
+import com.example.demo.exception.StudentNotFoundException;
 import com.example.demo.model.student;
+import com.example.demo.model.student;
+import com.example.demo.repository.StudentRepository;
 import com.example.demo.service.StudentService;
+
 
 @RestController
 public class StudentController {
 	 @Autowired
-	 private StudentService studentService;
-	 
+	 private StudentService studentService;	
 	 @RequestMapping("/create")
 	    public String create(@RequestParam String firstName, @RequestParam String lastName, @RequestParam int age,@RequestParam int sem, @RequestParam String sec,@RequestParam String usn) {
-	        student p = studentService.create(firstName, lastName, age,sem,sec,usn);
+		 if (firstName==null||lastName==null||((age==0)||(age<0))||sec==null||((sem==0)||(sem<0))||sec==null||usn==null)
+	        {
+			 throw new StudentNotFoundException("Incomplete information");
+	        }
+		 student p = studentService.create(firstName, lastName, age,sem,sec,usn);
 	        return p.toString();
-	    }
+	        
+	 }
+	 
 	    @PostMapping(value = "/")
 	    public ResponseEntity<?> saveOrUpdateStudent(@RequestBody student person) {
 	        studentService.createrecord(person);
 	        return new ResponseEntity("Student added successfully", HttpStatus.OK);
+	       /*
+	       if (firstName==null||lastName==null||((age==0)||(age<0))||sec==null||((sem==0)||(sem<0))||sec==null||usn==null)
+	        {
+				return new ResponseEntity("Incomplete information/wrong information ", HttpStatus.NOT_FOUND);
+	        }
+	        */
+	       
+	        
 	    }
 
 	    
@@ -45,10 +64,17 @@ public class StudentController {
 	    
 	    @GetMapping(value="/firstname/{firstName}")
 	    public student getstudentbyname(@PathVariable ("firstName") String firstName) {
+	    student s=studentService.getByFirstName(firstName);
+	    		if (s==null)
+			throw new StudentNotFoundException(firstName+" is not Found");
 	        return studentService.getByFirstName(firstName);
 	    }
 	    @GetMapping(value="/usn/{usn}")
-	    public student getstudentbyusn(@PathVariable ("usn") String usn) {
+	    public student  getstudentbyusn(@PathVariable ("usn") String usn) {
+	    	student s = StudentRepository.findByUsn(usn);
+	    	if (s==null)
+				throw new StudentNotFoundException(usn + " is not present");
+
 	        return studentService.getByUsn(usn);
 	    }
 	    
@@ -58,9 +84,15 @@ public class StudentController {
 	        return p.toString();
 	    }
 	    @PutMapping(value="/{usn}")
-	    public student updaterecord(@RequestBody student s, @PathVariable ("usn") String usn) {
-	        student p = studentService.updaterecord(s,usn);
-	        return p;
+	    public ResponseEntity<student>  updaterecord(@RequestBody student s, @PathVariable ("usn") String usn) {
+	    	student studentOptional = StudentRepository.findByUsn(usn);
+	    	if (studentOptional==null)
+				return new ResponseEntity("FAILED TO UPDATE ", HttpStatus.NOT_FOUND);
+
+
+	    	studentService.updaterecord(s,usn);
+	    	
+	    	return new ResponseEntity("UPDATED SUCCESSFULLY", HttpStatus.OK);
 	    }
 	    @RequestMapping("/delete")
 	    public String delete(@RequestParam String firstName) {
@@ -69,6 +101,9 @@ public class StudentController {
 	    }
 	    @DeleteMapping(value="/firstname/{firstName}")
 	    public String deleterecord(@PathVariable("firstName") String firstName){
+	    	student s = StudentRepository.findByFirstName(firstName);
+	    	if (s==null)
+				throw new StudentNotFoundException(  " is not present"+firstName);
 	        studentService.delete(firstName);
 	        return "deleted"+firstName;
 	    }
@@ -79,6 +114,9 @@ public class StudentController {
 	    }
 	    @DeleteMapping(value="/usn/{usn}")
 	    public String deletebyusn(@PathVariable("usn") String usn) {
+	    	student s = StudentRepository.findByUsn(usn);
+	    	if (s==null)
+				throw new StudentNotFoundException(usn + " is not present");
 	        studentService.deletebyusn(usn);
 	        return "deleted"+usn;
 	    }
@@ -87,5 +125,8 @@ public class StudentController {
 	        studentService.deleteAll();
 	        return "Deleted all records";
 	    }
-	   
+	 
+	 
+	
+	 
 }
